requirements.txt
torch
torchvision
tqdm
numpy
scikit-learn
matplotlib
pillow
flask
werkzeug
